# specs/001-web-ui-api/contracts/api-spec.yaml
openapi: 3.1.0
info:
  title: Ginkgo API Server
  description: Ginkgo量化交易系统API接口
  version: 1.0.0
  contact:
    name: Ginkgo Team
    email: support@ginkgo.trade
servers:
  - url: http://localhost:8000
    description: 开发环境
  - url: https://api.ginkgo.trade
    description: 生产环境
...
paths:
...
  /api/backtest:
    get:
      tags: [backtest]
      summary: 获取回测任务列表
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, RUNNING, COMPLETED, FAILED]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestListResponse'
    post:
      tags: [backtest]
      summary: 创建回测任务
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacktestCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResponse'
  /api/backtest/{uuid}/start:
    post:
      tags: [backtest]
      summary: 启动回测任务
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 启动成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BacktestCreate:
      type: object
      required:
        - name
        - start_date
        - end_date
        - initial_cash
        - components
      properties:
        name:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        initial_cash:
          type: number
          format: float
          minimum: 0
        components:
          type: object
          properties:
            strategy:
              type: string
            selector:
              type: string
            sizer:
              type: string
            risk_manager:
              type: string
    BacktestResponse:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        result:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
    BacktestListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BacktestResponse'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...